refactor(state-migration): move storage converters into internal files
